requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic>=2.10.0
aiofiles==23.2.1

# ML / Vision
torch>=2.6.0
torchvision>=0.21.0
timm>=1.0.0
Pillow>=11.0.0
numpy>=2.1.0
scikit-learn>=1.5.0
opencv-python-headless>=4.10.0.84
ImageHash==4.3.1
matplotlib>=3.9.0
